Time: 0 ms (100%), Space: 18.3 MB (91.66%) - LeetHub

+++ 0229-majority-element-ii/0229-majority-element-ii.cpp
@@ -7,7 +7,7 @@ public:
         int el2 = INT_MIN;
         int n=nums.size();
         
-        int q = (int)(n/3)+1;
+        int q = (n/3)+1;
 
         for(int i=0;i<n;i++){
             if (cnt1==0 && nums[i] != el2){
